requirements.txt
# Core dependencies
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytesseract>=0.3.10
python-Levenshtein>=0.21.0
wandb>=0.16.0
matplotlib>=3.7.0
torch>=2.0.0
open_clip_torch>=2.20.0

# Future dependencies (commented out for now)
# diffusers
# transformers
# surya-ocr
# deepseek-ocr
# groundingdino

